+++ alacritty/src/window_context.rs
@@ -328,6 +328,12 @@ impl WindowContext {
         self.dirty = true;
     }
 
+    /// Get reference to the window's configuration.
+    #[cfg(unix)]
+    pub fn config(&self) -> &UiConfig {
+        &self.config
+    }
+
     /// Clear the window config overrides.
     #[cfg(unix)]
     pub fn reset_window_config(&mut self, config: Rc<UiConfig>) {